test(backup): Add more thorough model coverage checks
We've already introduced model coverage checks in a previous PR, but
these only validate that every `__include_in_export = True`-marked model
is exercised by at least one test in
`tests/sentry/backup/test_models.py`. This change improves on that
slightly, by only considering a model to be "tested" if the matching
instance in the output has all of its fields set, and none of those
fields are set to their default value (if they have one). This ensures
that not only are we covering every exportable model, we are covering
every field of every such model.

Issue: getsentry/team-ospo#156